+++ tanks-game/_source/powers/mega-death-power.h
@@ -62,10 +62,10 @@ public:
 	virtual InteractionUpdateHolder<BulletUpdateStruct, WallUpdateStruct> modifiedCollisionWithWall(const Bullet*, const Wall*) override;
 	//bool modifiedCollisionWithWallCanWorkWithOthers = false;
 
-	virtual double getBulletSpeedMultiplier() const override { return 1.0/4; }
+	virtual float  getBulletSpeedMultiplier() const override { return 1.0/4; }
 	//with bulletSizeMultiplierPerTick = 257.0/256.0, 1.0/128 or 1.0/64 might be closer to what I originally thought of
 	//virtual double getBulletRadiusMultiplier() const override { return 1; }
-	virtual double getBulletAcceleration() const override { return 1.0/64; }
+	virtual float  getBulletAcceleration() const override { return 1.0/64; }
 	//with bulletSizeMultiplierPerTick = 257.0/256.0, 1.0/512 or 1.0/1024 (with a smaller radius multiplier per tick) might be closer to what I orignially thought of
 	virtual double getBulletRadiusGrowNumber_Stationary() const override { return MegaDeathPower::bulletSizeMultiplierPerTick; }
 	//bool bulletRadiusGrowMultiplies_Stationary = true;
