+++ tanks-game/_source/powers/mines-power.h
@@ -68,9 +68,9 @@ public:
 	virtual BulletPower* makeDuplicate() const override { return new MinesBulletPower(); }
 	virtual TankPower* makeTankPower() const override;
 
-	virtual double getBulletSpeedMultiplier() const override { return 0; }
+	virtual float getBulletSpeedMultiplier() const override { return 0; }
 	//bool bulletSpeedStacks = false; //true?
-	virtual double getBulletAcceleration() const override { return 0; }
+	virtual float getBulletAcceleration() const override { return 0; }
 	virtual float getBulletAccelerationImportance() const override { return .5; }
 
 	MinesBulletPower();
